pyproject.toml
[tool.poetry]
name = "wunder_decrypt"
version = "0.1.0"
description = ""
authors = ["laurihak <lauri_hakala@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
caesarcipher = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
